atari_jepa/requirements.txt
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
matplotlib>=3.8.0
pandas>=2.1.0
pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
tensorboard>=2.15.0
opencv-python>=4.8.0
einops>=0.7.0
